+++ inst/extdata/testdata/create-test-MultiGSEAResult.R
@@ -4,5 +4,5 @@ vm <- exampleExpressionSet(do.voom=TRUE)
 gsl <- exampleGeneSets()
 gsd <- conform(GeneSetDb(gsl), vm)
 x <- seas(gsd, vm, vm$design, methods=c('camera', 'fry'))
-path <- 'inst/extdata/testdata/test-MultiGSEAResult.rds'
+path <- 'inst/extdata/testdata/test-SparrowResult.rds'
 saveRDS(x, path)